@@ -1,16 +1,12 @@
 import Slider from "react-slick";
 import { Link } from "react-router-dom";
-import { useFetch, useConfigCarrousel } from "../hooks";
+import { useConfigCarrousel } from "../hooks";
 import Spinner from "./Spinner";
+import { useFetchSeriesQuery } from "../redux/api/apiSlice";
 import "../less/series.less";
-const apiKey = process.env.REACT_APP_API_KEY;
 
 const PosterSeries = () => {
-  const {
-    data: series,
-    isLoading,
-    error,
-  } = useFetch(`https://api.themoviedb.org/3/tv/popular?api_key=${apiKey}&language=es-ES`);
+  const { data: series, error, isLoading } = useFetchSeriesQuery();
 
   const { imageIndex, settings } = useConfigCarrousel();
 
@@ -27,7 +23,7 @@ const PosterSeries = () => {
       <h1>Series populares</h1>
       <div className='serie-container'>
         <Slider {...settings}>
-          {series?.map(({ id, name, poster_path }, index) => (
+          {series?.results?.map(({ id, name, poster_path }, index) => (
             <div key={id} className={index === imageIndex ? "slide4 activeSlide4" : "slide4"}>
               <Link to={`/serie/${id}`}>
                 <img src={`https://image.tmdb.org/t/p/w500${poster_path}`} alt={`${name}`} />
